mpris: Improve metadata extra field support.
[mpris] Improve metadata extra field support. JB#59021

The initial extraFields implementation omitted mpris: and xesam: ontologies
though the latter is permissible and even recommended in the spec. Switch
this to only ignore the fields that are used elsewhere.

Add extraField companion methods for settings individual fields for
easier use.

Plus a small tweak to allow fields to be unset properly only when required.
This was causing null values to be set repeatedly, and metadataChanged
signals to be sent out too often.