mpris: Store enum values as integers in MprisPlayerInterface
[mpris] Store enum values as integers in MprisPlayerInterface

Fix invalid initialization in MprisPlayerInterface, by storing enum
values as integers instead of strings. Previously enum values were
stored as strings in MprisPlayerInterface and converted elsewhere in the
code. Now they are stored as enum values and converted when accessed
through accessors.

Note that this will change the behaviour a little as now setting an
invalid value does not result in change at all. Previously it would
have set the value as is and made the library to effectively use a
default value. Internal functions access the values directly through
internal getters and setters.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>