mpris: Validate trackId as object path
[mpris] Validate trackId as object path. JB#59304

Validate that trackId is a valid object path before setting it. If it
happens to be QDBusObjectPath already, then convert it to string
properly. Null QVariant is passed through as it becomes NoTrack path
once converted to QDBusObjectPath when accessed over D-Bus. Invalid
values print a warning and set trackId to null.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>