Adjust overloaded parent() handling to named member variables
parent() is conventionally the owning QObject instance in Qt.
Here it being used to that but also other needs was making it hard
to follow the code. Was often mistaking references to private as
references to owning object of the public instance.

And then there were strange setups like parent()->parent()->something(),
where the first parent was reference to another class instance private
and latter parent() to its public class.

Mostly mechanic changes except MprisPlayerAdaptor::minimumRate()
wasn't following the pattern for some reason.