[mpris] Fix qml api problem for following metadata changes. JB#58070
Since the MprisController metaData instance was changing on different
apps to follow, qml code such as

MprisController {
  metaData.onTitleChanged: console.log(metadata.title)
}

...didn't work. The connection was created for the initial metaData
instance, not updated later on.

Fixed by a separate proxy class, keeping the returned instance of
MprisController always the same. The result is a bit verbose at the
moment but it's simple. On longer run we should probably better separate
the metadata instances on media and controller apps.