[mpris] Use enums internally for LoopStatus and playbackStatus. Fixes JB#59562
Rather than storing the loopStatus and playbackStatus as QStrings, this
uses the enums internally to ensure valid values are always maintained.

Note that this will change the behaviour a little as now setting an
invalid value does not result in change at all. Previously it would
have set the value as is and made the library effectively use a default
value. Internal functions access the values directly through internal
getters and setters.

Co-authored-by: Tomi Leppänen <tomi.leppanen@jolla.com>